+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.5.5:
+* Introduce ReplicantContextHolder to ease passing of context information between
+  tiers.
+
 ## 0.5.4:
 * Import BadSessionException from downstream projects.
 * Add synchronized modifier to several methods of PacketQueue.